Game engine beginnings
Imported engine code from insanity sweeper and cleaned it up a bit,
adding comments where necessary. I will clean up this code more and
comment more when I have the chance (see added todos for some
improvements that can be made). Hopefully armed with my new knowledge
on data structures and programming from university I will be able to
create an engine that makes sense to more people than just me. I aim to
make the engine less dependent on you implementing things directly into
the render engine and more dependent on just adding shaders and textures
to an entity and slapping it away to the render engine to be drawn.
NOTE: For now the render engine is OpenGL 1.1+ but in the future it will
require 2.0+ (with a seperate renderer for 3.3+) however I want to get
base functionality down and some prototype classes for entities and
shaders done before that.